requirements.txt
openai>=1.3.0
requests>=2.25.1
python-dotenv>=0.19.0
Pillow>=8.3.2 